refactoring: better interface segregation
